depthcharge: Specify where the kernel loading area is.
The kernel needs to be loaded into an area backed by RAM. Also make sure that
the kernel loading area and the base of depthcharge have the same relative
order that they do in the linker script.

BUG=chrome-os-partner:17340
TEST=Built and booted into depthcharge on Snow. With this change, Coreboot no
longer complained about loading part of depthcharge into an area that isn't
memory.
BRANCH=None

Change-Id: I7c157a0624758362140eb9f7b5db2c4dfc8e2728
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/32954
Reviewed-by: Stefan Reinauer <reinauer@google.com>