depthcharge: Add a trampoline entry point for ARM.
Without an entry point, the trampoline code was being garbage collected by the
linker entirely, and when adding the empty region it would occupy to the code
that tracks what memory is unused it triggered an assert. This code is not
functionally tested yet.

BUG=chrome-os-partner:17340
TEST=Built and booted into depthcharge on Snow. With this change, saw an
assert about an empty range go away. Checked the sections in the depthcharge
ELF before and after this change. There were .tramp sections afterwords but
not before.
BRANCH=None

Change-Id: Ica110558adb64c379bad2f01a63e92c2b5780f2d
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/32955
Reviewed-by: Stefan Reinauer <reinauer@google.com>