depthcharge: Update the LPC TPM driver and the vboot callbacks.
In U-Boot, the LPC TPM driver has changed, and more importantly the functions
it provides to the vboot callbacks have changed, simplifying the callbacks
presumably by moving controlling logic into the driver itself. In general the
change seems to be for the better although I haven't looked into it that
deeply, but it's also necessary in order to interface with the I2C TPM driver
currently in U-Boot.

BUG=chrome-os-partner:17340
TEST=Built and booted depthcharge on Link. Looked at the log and verified that
there were no TPM related errors and that TPM operations were happening and
seemed to be succeeding. Verified that performance with the new driver was
approximately equivalent to the old one.
BRANCH=None

Change-Id: I31c1d4ad068a4110d021f2f24aff0babe6bef5a6
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/32996
Reviewed-by: Stefan Reinauer <reinauer@google.com>