depthcharge: In the I2C driver, set up I2C bus 3.
There are pins which need to be configured so that they're attached to I2C bus
3. That setup should ultimately not be in the I2C driver itself, but other
similar pins are already being set up there, and there isn't a better place at
the moment.

BUG=chrome-os-partner:17340
TEST=Built and booted into depthcharge on Snow and saw some reaction from the
driver when attempting to use I2C bus 3. Previously it would time out waiting
for an interrupt.
BRANCH=None

Change-Id: I3e92d88ea08a51e6ca69bbfab551f50ef3cde4c1
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/32998