depthcharge: fix RW firmware hashing size
All depthcharge device trees use the 'with_index' property for the
RW sections of the flash. The presence of this property in the
device tree adds a book keeping header with the number of components
and each component's size and offset. Therefore, with determining
the size of the data to hash one needs to always use the book keeping
header to calculate the correct size.

BUG=None
BRANCH=None
TEST=Can properly boot the RW path on haswell systems which contain a
     single firmware compoment.

Change-Id: I32b1f96a7414ca867b88b30a02eaf8eae2fce0f5
Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://gerrit-int.chromium.org/33049
Reviewed-by: Duncan Laurie <dlaurie@google.com>