depthcharge: Fix a few minor issues with the exynos SPI driver.
This change fixes up four things in the exynos SPI driver.

1. The packet count when sending the command/offset was set to the size of a
uint32_t which is the number of bytes to send, but packets are 4 bytes long
and the count should have been 1.

2. When receiving data from the RX FIFO, we weren't decrementing the "words"
variable which keeps track of how many words we're still expecting. This
variable can limit the amount of stuff put in the TX FIFO.

3. When doing a software reset, the manual suggests turning off the RX and TX
channels first. This change introduces a function which turns off the
channels, does the reset, and then re-enables the channels. This function is
used before changing the packet count and while setting up the channel at the
start of the transaction.

4. The "high speed enable" bit of the CH_CFG register is cleared instead of
being set. This matches what the U-Boot driver does, but since the manual
suggests this only matters when transmitting in slave mode (which we never do)
it's not clear if it actually matters. I change it anyway just in case.

BUG=chrome-os-partner:17340
TEST=Without this change, there's a non-determinstic hang when booting on
Snow. With too much output the hang stops happening, but I'm fairly confident
it's happening in the SPI driver. With this change and no additional output,
my Snow has rebooted itself over and over for hours without encountering the
hang. I also haven't seen any SPI related error messages or failures.
BRANCH=None

Change-Id: Ifa3750befe702bf48f3fbfe7821541408bba2074
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/33219
Reviewed-by: Ronald G. Minnich <rminnich@google.com>