depthcharge: Don't print an error when the chip number is nacked in s3c24x0.
The TPM is apparently very sleepy and/or unresponsive, and will nack
frequently even when it's working correctly. The TPM driver seems to
compensate for this by sending accesses multiple times. To avoid lots of junk
error messages, this change removes the printf which reports a nack when
sending the chip number on the s3c24x0 i2c bus. If a data packet is nacked,
however, an error is still printed.

BUG=chrome-os-partner:17340
TEST=Built and booted depthcharge on Snow with an I2C TPM driver in place. Saw
spurious error messages from the I2C driver go away.
BRANCH=None

Change-Id: I3346b7ad6ebabfc1a2a27731e21b6bd61406dcc1
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/33222
Reviewed-by: Ronald G. Minnich <rminnich@google.com>