depthcharge: Port over the U-Boot I2C based TPM driver.
This driver is taken mostly as is from U-Boot, but with some things stripped
out, specifically device tree support and Linux error codes. The error codes
have been left intact but are commented out in case we want to distinguish
between types of errors at some point in the future.

This driver also seems to be a bit too spread out and sprawling for what it's
supposed to be doing. I'm not very familiar with how TPMs or TPMs over i2c
are supposed to work, though, so I can't say for sure or do much about it.

BUG=chrome-os-partner:17340
TEST=Built and booted into depthcharge on Snow with this I2C TPM driver turned
on. Saw that vboot was able to initialize it successfully and get the
information it wanted out of it.
BRANCH=None

Change-Id: Ia576fad48cd7b4fe8a179c926633b594450fe45e
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/33223
Reviewed-by: Ronald G. Minnich <rminnich@google.com>