depthcharge: Disable rollback detection on Daisy.
MP snows have their TPMs rolled formward to firmware version 2, but firmware
built from ToT, including U-Boot, signs itself as version 1. There are three
possible solutions to this problem.

1. Roll the TPM back. This is possible but very annoying for even a single
developer. Having to do this for everyone that wants to run depthcharge on
their snow would be prohibitive.

2. Sign the firmware with version 2. This is tempting, but I suspect we didn't
do this on other platforms for some reason. It's like that we conciously chose
for MP firmware to be version 2 and for developer built firmware to be version
1. If there isn't really a good reason not to go this way, it would be nice.

3. Disable the rollback check all together. This is the solution everything
based on U-Boot (which is essentially everything right now) uses. It makes me
a little uneasy to disable one of our security mechanisms and make it so it's
not being exercised regularly, but this is consistent with other platforms and
may have some compelling advantage I'm not aware of.

I chose option 3.

BUG=chrome-os-partner:17340
TEST=Built and booted into depthcharge on Snow and saw TPM rollback errors go
away. vboot is now happy with the first RW firmware.
BRANCH=None

Change-Id: I16646895a189e3d258c51e11b9883ba5dc152b17
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/33225
Reviewed-by: Ronald G. Minnich <rminnich@google.com>