depthcharge: If VbSelectAndLoadKernel requests powering down, do that.
Without special handling, the default is to reboot. That's not the intended
behavior when the error code is VBERROR_SHUTDOWN_REQUESTED, however.
Unfortunately the comments in vboot_api.h don't say anything about handling
this code specially, but that's what U-Boot is doing so it's likely correct.

BUG=chrome-os-partner:8339
TEST=Built and booted into depthcharge on Snow and Link with the lid switch
forced to be closed and actually closed, respectively, and saw that the
machine powered off instead of rebooting.
BRANCH=None

Change-Id: Id4f6abcd26bd7d14c6480e4a37bc12d72774b3e0
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/33226
Reviewed-by: Bill Richardson <wfrichar@google.com>