depthcharge: Rework the Exynos5 I2C driver to work with the Maxim codec.
The maxim codec expects I2C transactions to be in a particular format which
the I2C driver wasn't using. In general expecting I2C devices to use a
particular structure to the data it sends and receives is doomed to failure,
but the current API tends to be convenient. In the future, it would be a good
idea to rework the API to more reflect the way I2C actually works and then to
build some convenience functions on top of it that make it easier to use the
API in common ways.

BUG=chrome-os-partner:17340
TEST=Built and booted into depthcharge on Daisy and verified that vboot could
still talk to the EC and TPM successfully. With other changes, configured the
Maxim codec successfully over I2C.
BRANCH=None

Change-Id: I721a8fa70e3773df11a8bed91296e2b800789ec1
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/33363