depthcharge: Add support for I2S based sound and the Maxim98095 codec.
These were both copied out of U-Boot and reworked a bit. Also, the config
option for turning on the HDA codec driver was renamed to be consistent with
the other options, and the dummy driver is no longer the default for ARM and
has to be configured explicitly.

BUG=chrome-os-partner:17340
TEST=Built and booted into depthcharge on daisy. With this driver turned on
and code added to ro_main to call it, made the speakers beep.
BRANCH=None

Change-Id: I2a2f759cffe67364dc6bdb4053e367e3e0efff97
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/33364