depthcharge: Rework the bitmap display code to avoid unaligned accesses.
A lot of the metadata in the bitmap format is inherently unaligned and causes
data aborts when accessed on Daisy. This change reworks the bitmap display
code so that it copies data to addresses that make things aligned.

BUG=chrome-os-partner:17340
TEST=Built and booted into dev mode on Daisy and saw some data aborts go away.
Built and booted into dev mode on Link and verified that bitmaps were
decompressed/displayed correctly.
BRANCH=None

Change-Id: I41b878e2dea11ae23127d248e65d158bc8470a25
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit-int.chromium.org/33368