depthcharge: Get rid of some unused constants in the Exynos I2C driver.
A future CL is going to change the endian conversion macros into inline
functions, and the definition of these constants upsets the compiler because
they'd become the product of function calls, not expressions that can be
evaluated directly. Since they aren't even used, they can simply be deleted.

BUG=chrome-os-partner:8339
TEST=Built for Daisy.
BRANCH=None

Change-Id: I400401284b639c6afc6a6300ed4f8999ffc7e8b2
Signed-off-by: Gabe Black <gabeblack@google.com>