depthcharge: Move some common booting code into vboot/stages.c.
Figuring out where the various bits of a vboot kernel blob are doesn't depend
on what format the kernel part of the blob is in. That code can be in the
common vboot/stages.c and shared. It's also nice to move this vboot specific
code out of the boot directory and into the vboot directory.

The actual calculation of the location of the various components has also been
improved to account for the fact that some kernels claim they want to be
loaded at 1MB, even though that's sometimes not even possible.

BUG=chrome-os-partner:8339
TEST=Built and booted on Link.
BRANCH=None

Change-Id: Ic49daac4c50287f3c81811433c45341721f463f1
Signed-off-by: Gabe Black <gabeblack@google.com>