depthcharge: Add some code for manipulating device trees.
This code makes it fairly easy to read and manipulate device trees without
being too big or complicated. This code is for managing the kernel's device
tree only, not as a precursor of device tree support in depthcharge.

BUG=chrome-os-partner:17340
TEST=Built and booted into depthcharge on Snow with some other CLs that add
support for the FIT image type and verified that the image could be read
correctly and that the kernel's device tree could be read and modified
correctly.
BRANCH=None

Change-Id: I9e23f89976b8769fb1d56334cadd6c8e9acad438
Signed-off-by: Gabe Black <gabeblack@google.com>