depthcharge: Rearrange the kernel boot code.
Move around the code that boots the kernel yet again. The concept of a zimage
is not unique to x86, and also describes the way the kernel is packaged, not
the way it's booted. The zero page structure is actually the boot mechanism
for x86 no matter how the kernel was packaged on disk.

To seperate the ideas of the kernel's boot protocol on each architecture and
various ways of packaging a kernel, the zimage code was moved into arch/x86
and renamed generically to boot.c boot.h, and a collection of support
files in arch/x86/boot/. The actual zimage format is now supported by having a
"boot" function in zimage.c which can be selectively compiled in. It takes the
kernel elements which were returned by vboot, interprets their contents as a
zimage, and passes them on to the x86 specific boot protocol.

Because the zimage format doesn't require any additional processing other than
what's required by the actual boot protocol, the zimage version of boot
doesn't do anything other than just adapt one function call to another, plus
updating the crossstem data. A different format, for instance FIT on ARM,
would require more work before the kernel was ready to boot.

BUG=chrome-os-partner:17340
TEST=Built and booted on Link. Built and booted into depthcharge on Snow.
BRANCH=None

Change-Id: I0c1649364cac34afbf10807e2a4bfd2964752fc3
Signed-off-by: Gabe Black <gabeblack@google.com>