depthcharge: Add some generic code for handling FIT images.
This code unpacks and prepares a FIT image and chooses what config to boot.
It's first preference is to boot the config who's FDT's compatibility property
matches a Kconfig option. If that fails, it will choose the default. If even
that fails, it will give up and return an error.

When filling in the "memory" node, this code will try to determine where RAM
is and what areas need to be marked as reserved in the reserve memory map
passed through the device tree.

BUG=chrome-os-partner:17340
TEST=With this and another change, built and booted into the beginning of the
kernel on Snow.
BRANCH=None

Change-Id: I8ec87b7cc6c28b74d8de150a284ceb8a659e3c9c
Signed-off-by: Gabe Black <gabeblack@google.com>