depthcharge: Update device tree for fox boards
The fox boards will be supporting a signed ramstage. Therefore
add this blob type to the device tree for fox boards. Also, clear
the preamble-flags in the vblock because the RO normal path
is not a supported option on fox boards. The verified firmware is
always used to boot unless recovery mode is enacted.

BUG=chrome-os-partner:17695
BRANCH=None
TEST=Built with unified boot loader and ebuild changes. Can verify
     ramstage and load properly.

Change-Id: I912d532d1d71d5529774aa22de8316204d53a247
Signed-off-by: Aaron Durbin <adurbin@chromium.org>