Make sure the command line and boot_params out of the way on x86.
When booting on x86, we build the command line within depthcharge which is
passed in place to the kernel as it starts. When the kernel decompresses
itself, there isn't anything that prevents it from expanding upwards until it
overwrites the command line. That hasn't happened yet because the kernel
hasn't been big enough to cover that gap. The boot_params are copied to
somewhere safe by the kernel early in boot and are generally less of a
problem. In the kernel's boot protocol documentation, it says that the
bootloader should be loaded at 0x1000 and that the command line and
boot_params should be put at that plus "X", where X is presumeably the size of
the bootloader itself. Since we're not loaded down there, we can just let X be
zero.

BUG=chrome-os-partner:16688
TEST=Built and booted on Link.
BRANCH=None

Change-Id: I26c319183f24e1f448b9940e5524364e9f7665f0
Signed-off-by: Gabe Black <gabeblack@google.com>