Figure out the firmware ID without depending on the ACPI data.
The ACPI data may not have been updated by the time the debug info is
requested, and it's not even present on ARM. This change makes the display
callback figure out which id to use on its own in the same way as the ACPI
code would.

BUG=chrome-os-partner:17340
TEST=Built and booted into RO and RW versions of depthcharge on Link, pressed
TAB, and verified that firmware IDs were displayed properly.
BRANCH=None

Change-Id: I9f38d7b3dd35e1878f56c80acb6e9944ff21b518
Signed-off-by: Gabe Black <gabeblack@google.com>