If the boot_params haven't been separated from the kernel, do that manually.
When booting a zimage with vboot, the boot_params structure has already been
seperated out. When booting one otherwise, it has to be seperated out
manually. If the params pointer is NULL, then we should assume the boot_params
are still stuck to the front of the kernel.

BUG=chrome-os-partner:16688
TEST=Built and did a netboot on Link.
BRANCH=None

Change-Id: Id28a20f6eb00fceda2503b9feca744db03748390
Signed-off-by: Gabe Black <gabeblack@google.com>