netboot: Add facilities for setting netboot parameters after an image is built.
The parameters are put in the SHARED_DATA section of the image which is not
currently used by depthcharge at all, and would likely only be used when
running the normal, non-netbooting version.

There is a string "netboot" with a NULL terminator at the start of the data
which signifies that there's something at least partially reasonable and
intended for use by netboot in the SHARED_DATA section.

After that is a 32 bit, little endian count which says how many parameters
follow.

For each parameter, there's first a 32 bit, little endian tag which describes
which parameter is being set. Next, there's a 32 bit, little endian tag which
is the length of the parameter in bytes, not including the type or length
fields. The data follows, and is padded out to be a multiple of 4 bytes long.

The currently recognized parameter tags are:

1: TFTP server IP address as a 4 bytes in network byte order.
2: Additional command line arguments.
3: The file to download over TFTP and boot.

BUG=chrome-os-partner:16688
TEST=Built with dynamic parameters (the old behavior) turned on, and saw
netboot work when those parameters were provided by the DHCP server. Built
with static only parameters turned on, but none set, and saw an appropriate
failure message. Booted with those parameters set and saw a successful netboot.
Booted while plugged into the corporate network and saw that DHCP finished
successfully. The following TFTP download failed with a "file not found" error,
presumably because it was redirected to different TFTP server.
BRANCH=None

Change-Id: I170af37d8cb8fe120e7f959577af465dd219fc06
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/45944
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>