Clean up the Makefile related to unified depthcharge.
The conditional behavior can be isolated to where real-target is defined. We
can append new dependencies conditionally and keep the original definition
unmodified.

Also, the actual rule that builds the payload version of the unified image has
been split up so that it runs as seperate commands. That will make it easier
to tell what happened if a command in the middle fails. A message has been
added which says what the Makefile is currently doing if the V flag (verbose
mode) hasn't been turned on.

BUG=None
TEST=Built for fox, link, daisy.
BRANCH=None

Change-Id: I093cd48f97a4c7ceb9860b6721292340fa4abb04
Signed-off-by: Gabe Black <gabeblack@google.com>