uip: Get rid of some code that could never be compiled in.
The entire uip.c file was guarded so that it wouldn't be compiled when
building for IPv6, but then individual areas within the file were guarded so
that they would only be compiled in if IPv6 was on, or redundantally checked
that it wasn't. The code that could never be compiled in was eliminated
entirely, as were the redundant checks.

BUG=None
TEST=Built and netbooted on Link.
BRANCH=None

Change-Id: I2e2441f74f9cf636ef368edc06d77974d7c27e1c
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/46855
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>