Use $(MAKE) instead of make so vboot is compiled with -j.
When running make recursively, make will automatically pass through some
options if it can tell it's running recursively, specifically if it's invoked
with $(MAKE) or ${MAKE}. When it is, the top level make and the sub-makes
coordinate with each other so that the -j option of the top level make is
honored and that the appropriate number of jobs run in parallel. Without that,
the submake defaults back to -j1 which can be much slower.

BUG=None
TEST=Built and booted on Link. With this change, a warning from make about
running with -j1 went away, and the recursive make of vboot ran significantly
faster.
BRANCH=None

Change-Id: If601076670640faaa401f9c5f1e4567888fe1a7e
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/46858
Reviewed-by: Bill Richardson <wfrichar@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>