uip: Strip out IPv6 support.
The only aspects of uIP's IPv6 support we actually had were bits which were
alongside code we needed for IPv4. Since we couldn't support it as is if we
wanted to, there's no reason to have to those bits floating around getting in
the way.

BUG=None
TEST=Built and netbooted on Link.
BRANCH=None

Change-Id: I22037f1fc3a4414308d80c7bb3719479b3cd9c24
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/46999
Reviewed-by: Julius Werner <jwerner@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>