Ask the kernel not to reload it's code/data segments early in boot.
The kernel will reload its code and data segments early in boot unless it's
asked not to. Doing so requires that the GDT have particular segments in
particular places, and that it's still accessible once the kernel starts. This
has historically been true because the GDT came from Coreboot and Coreboot has
been out of the way. All the kernel really wants, though, is to ensure that
the code and data segments are flat 4GB, and we're already providing that. We
can pass in the KEEP_SEGMENTS flag to tell Linux to just use the segments we
gave it until it can set up its own GDT.

BUG=chrome-os-partner:18390
TEST=Built and booted on Link.
BRANCH=None

Change-Id: Ie53e0c08214185e33155f1d06f85328c59b3523b
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/47273
Reviewed-by: Stefan Reinauer <reinauer@google.com>
Reviewed-by: Ronald G. Minnich <rminnich@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>