ARM: Add the crossystem data to the kernel's device tree when booting on ARM.
This functionality was currently stubbed out and not implemented. Several
values that are included in U-Boot are excluded here because the crossystem
utility doesn't look at them.

The ddr-type property is not set despite the fact that U-Boot sets it and
crossystem reads it. There are two main reasons for that. First, the existence
of the value is board specific. Support for it should go in a board specific
place and not in the generic code added here.

Second, there's no reason for the firmware to have to prepare this value. The
different ddr types are selected based on whether certain resistors are
stuffed. Since the very existence of this field is system specific, similarly
system specific code could be added to the kernel, to crossystem, or to some
unrelated utility to read the GPIOs those resistors are connected to and
determine the ddr type.

BUG=chrome-os-partner:18639
TEST=Built and booted into ChromeOS on Snow. Ran crossystem and verified that
the same values were successfully retrieved as under U-Boot. Verified that
most values made sense, or at least were being passed through to crossystem in
a sensible way. One value in particular, the write protect value read by
coreboot, seems to be wrong. I believe that's because Coreboot is reporting it
incorrectly.
BRANCH=None

Change-Id: I870aeeec15e82b59851d41a95e6ba9abffab38f6
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/47542
Reviewed-by: Ronald G. Minnich <rminnich@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>