Add a "driver" which adds a cleanup function to disable video DMAing.
If the video hardware is still DMAing memory when the kernel starts, it can
(and currently does) enable the sysmmu without turning it off or configuring
the sysmmu to accomodate those transfers which results in a kernel panic. This
change adds some code which can optionally be compiled in which will turn off
the DMAing right before we turn things over to the kernel.

BUG=chrome-os-partner:18634
TEST=With this "driver" enabled, booted on Daisy and saw the kernel panic go
away.
BRANCH=None

Change-Id: Ibb794a3b24e66d774b550a31923e915d6223f534
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/48772
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>