Update EC commands header
The EC codebase has made a number of changes to the common
shared header file that are incorporated here.  The only
affected part in depthcharge was the rename of the flash
protect constants.

The header file is copied in exactly as it is in the EC
codebase so it can be compared easily to see if it needs
updating in the future.  (ideally this file would not
be renamed either)

BUG=chrome-os-partner:19874
BRANCH=none
TEST=manual: This was tested on a slippy board

- test normal mode boot
- test recovery mode boot
- test recovery->developer transition
- test developer mode boot
- test developer->normal transition
- test EC software sync

Change-Id: I210dead72880e0b552c763a27981c414fda8b504
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/57226
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>