netboot: Built RW binary and payload formated netboot blobs.
With early firmware selection turned on, we need to be able to handle
booting into RW or RO firmware even if we're not going to run vboot within
depthcharge. To that end, we need to build netboot blobs we can install as RW
firmware.

BUG=chrome-os-partner:20392
TEST=Built and booted into RO normal and netboot images. Can't do an RW boot
because of a potential bug in the EC.
BRANCH=None

Change-Id: Ic5f4c3603777f80921c53910ab9cc49d164813e5
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/59656
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>