Add board specific directories/files.
The idea is that the board.c in each of these directories will have an init
func in it which will instantiate and interconnect driver instances which fit
with that particular board. That will make things more flexible for drivers,
and avoid having to have hacks and board specifics wedged into drivers to
support specific boards.

For instance, on snow, one of the I2C busses is shared with the EC, and is
arbitrated using a request/grant pair of GPIOs. Instead of having the driver
know about that arbitration mechanism and which bus it goes with and which
GPIOs to use, the GPIOs could be instantiated as instances of the exynos GPIO
driver, the I2C driver can drive the most busses directly, and a wrapper which
handles the arbitration can be implemented which is called into by the device
driver on the I2C bus and which, after arbitration, then calls into the
generic driver.

Another example would be if there are two different types of I2C busses like
on both Exynos chips. There actually are two types of busses on both those
chips, but we've been able to ignore that by either turning off one type, or
only attaching devices to one type. If we can't or don't want to work around
the problem any more, we could implement drivers for both types of controller
and then pass the right instance to the right driver for the devices on those
busses.

BUG=None
TEST=Built for all supported boards. Booted on Link.
BRANCH=None

Change-Id: I349ebf4df1fc8d5763aa7d49fd70c85cc333c9f3
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/60983
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>