mkbp_read_hash: Ask EC to calculate EC_RW hash if unavailable.
If EC_RW hash is not available (ex. if we just reflashed the EC),
mkbp_read_hash will immediately fail. Instead, we should handle
unavailable EC hash more robustly -- if hash calculation is in progress,
wait for it to complete. If hash calculation has not started, try to
start it, and wait for it to complete. This will prevent us from
incorrectly entering recovery mode in some cases when we do an EC reflash.

Note: This implementation is very close to the existing u-boot function
cros_ec_read_hash.

BUG=chrome-os-partner:20809.
TEST=Manual. Reflash EC w/ flashrom and reboot. Observe "No valid
hash... Compute one" message on debug terminal, resulting in successful
boot.
Signed-off-by: Shawn Nematbakhsh <shawnn@chromium.org>

Change-Id: I28d5865ec4dddf0837e99d486019a3dc9753e835
Reviewed-on: https://gerrit.chromium.org/gerrit/61622
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Shawn Nematbakhsh <shawnn@chromium.org>
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Commit-Queue: Shawn Nematbakhsh <shawnn@chromium.org>