pit: Adjust the fixed addresses for various bits and pieces.
The part of the address space assigned to RAM starts at 512 MB instead of 1 GB
like on the Exynos5250. This change adjust various fixed addresses in the
config down by 512 MB so that their location relative to the start of memory
is the same.

BUG=chrome-os-partner:19420
TEST=Built and booted into the kernel on pit. Saw that the kernel no longer
crashes when trying to set up paging.
BRANCH=None

Change-Id: I5732530650d6446c0b9a234e541c772a14ad2391
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/61818
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>