WIP: DHCP: Wait for a response for 10 seconds instead of the polling delay.
Instead of assuming the amount of time spend waiting for the network device
to respond is an adequate timeout, wait for a fixed amount of time, currently
set to 10 seconds, before resending the request.

This code is untested other than making sure it compiles.

BUG=chrome-os-partner:20861
TEST=Compiled for falco.
BRANCH=None

Change-Id: If99a936464e74977cf20f605c26642797246b7ae
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/62219
Commit-Queue: Bowgo Tsai <bowgotsai@chromium.org>
Tested-by: Bowgo Tsai <bowgotsai@chromium.org>
Reviewed-by: Gabe Black <gabeblack@chromium.org>